Struct curve41417::ed::GroupElem

pub struct GroupElem {
    // some fields omitted
}

A group element i.e. a point in Edwards representation.

It handle various group elements operations such as scalar multiplications, point additions, Elligator point mapping, key pairs generation,...

Methods

impl GroupElem

fn new() -> GroupElem

Return a new group element with all its coordinates set to zero.

fn zero() -> GroupElem

Return a group element with all its coordinates set to zero.

fn neutral() -> GroupElem

Return the neutral point.

fn base() -> GroupElem

Return the base point (x, 34) of prime order L.

fn gen_key() -> ProtKey8

Generate a new secret key at random. It is clamped before being returned.

fn pack(&self) -> ProtBuf8

Pack a group elem's coordinate y along with a sign bit taken from its x coordinate. This packed point may later be unpacked with unpack().

fn unpack<T: AsRef<[u8]>>(bytes: &T) -> Option<GroupElem>

Unpack a Curve41417 point in Edwards representation from its bytes representation. bytes must hold a point packed obtained from a previous call to pack().

fn pack_to_mont(&self) -> ProtBuf8

Convert and pack this point from Edwards coordinates to Montgomery's x-coordinate. This result may be used as input point in curve41417::mont scalar multiplications.

fn unpack_from_mont<T: AsRef<[u8]>>(bytes: &T, sign_bit: u8) -> Option<GroupElem>

Unpack point from packed Montgomery's x-coordinate along with a sign bit originally obtained from the method sign_bit(). This unpacking may be used in a similar case than this, see also this discussion.

fn sign_bit(&self) -> u8

Return the sign bit, a value equal to zero or one exclusively, associated with the packed representation of this point. This returned value may then be used as argument when calling unpack_from_mont().

fn scalar_mult_base<T: AsRef<[u8]>>(n: &T) -> GroupElem

Return a point q such that q=n.BP where n is a scalar value applied to the base point BP. Note that n is not clamped by this method before the multiplication is performed. Calling this method is equivalent to calling GroupElem::base().scalar_mult(&n).

fn double_scalar_mult<T: AsRef<[u8]>>(n1: &T, p1: &GroupElem, n2: &T, p2: &GroupElem) -> GroupElem

Return a point q such that q=n1.p1+n2.p2 where n1 and n2 are scalar values and p1 and p2 are group elements. Note that the values of n1 and n2 are not clamped by this method before their respective multiplications.

fn scalar_mult<T: AsRef<[u8]>>(&self, n: &T) -> GroupElem

Return a point q such that q=n.self where n is the scalar value applied to the point self. Note that the value of n is not clamped by this method before the scalar multiplication is performed. Especially no cofactor multiplication is implicitly applied. Use scalar_mult_cofactor() prior to calling this method to explictly pre-cofactor this point.

This method deliberately takes as input parameter a raw scalar instead of a ScalarElem for the reason that in some cases we don't want the bits of the scalar to be modified before any scalar multiplication. Whereas a ScalarElem would automatically be reduced mod L (see base()) before any scalar multiplication takes place.

fn scalar_mult_cofactor(&self) -> GroupElem

Return a point q such that q=8.self where 8 is curve's cofactor applied to this point's instance.

fn elligator_to_representation(&self) -> Option<ProtBuf8>

Use Elligator to encode a curve point to a byte-string representation.

fn elligator_from_representation<T: AsRef<[u8]>>(r: &T) -> Option<GroupElem>

Map a byte-string representation to a curve point. Return a valid group element if it produces to a well-defined point.

The argument provided to this method is expected to originate from the encoded representation returned by elligator_to_representation(). Otherwise use elligator_from_bytes() instead.

fn elligator_from_bytes<T: AsRef<[u8]>>(s: &T) -> Option<GroupElem>

Map a byte-string to a curve point. Return a valid group element if s produces to a well-defined point. The provided input is first reduced in Fq. For better uniformity of the distribution in Fq the input must be sufficiently large i.e. between [64, 104] bytes.

Note that since ψ(r) = ψ(-r) with ψ the mapping function and r = s mod q, as r and -r map to the same point, it may lead to obtain a non unique result for elligator_to_representation(elligator_from_bytes(s)) in {r, -r}.