Function curve41417::mont::scalar_mult

pub fn scalar_mult<T: AsRef<[u8]>, U: AsRef<[u8]>>(n: &T, p: &U) -> ProtBuf8

Compute scalar multiplication

Return a packed point q such that q=n.p, n is a scalar and p is a point. On input, n is clamped before performing the scalar multiplication.