Struct curve41417::sc::ScalarElem

pub struct ScalarElem {
    // some fields omitted
}

Scalar element used in scalar operations.

Provide commons Curve41417 scalar operations computed mod L, where L is the prime order of the base point ed::base() currently used.

Methods

impl ScalarElem

fn new_rand() -> ScalarElem

Generate a new random ScalarElem between [0, L-1]. Use OsRng (which uses /dev/[u]random on Unix) as PRNG.

fn zero() -> ScalarElem

Return scalar value representing zero.

fn len(&self) -> usize

fn pack(&self) -> ProtBuf8

Pack the current scalar value reduced mod L. Note: it is not until this method is called that the scalar element is reduced to its canonical form.

fn is_zero(&self) -> bool

Check if self is equal to 0.

fn unpack<T: AsRef<[u8]>>(n: &T) -> Option<ScalarElem>

Unpack scalar value n. It must represent a value strictly in [0, L-1] and it should not be expected to be reduced on unpacking.

fn unpack_from_bytes<T: AsRef<[u8]>>(b: &T) -> Option<ScalarElem>

Unpack bytes b as a scalar and reduce it mod L. It is expected to be of a large enough size in order to provide a good uniformity of distribution on reductions mod L. Thus b must be between [64, 104] bytes.

fn reduce_from_bytes<T: AsRef<[u8]>>(b: &T) -> ProtBuf8

Unpack bytes b as a scalar, reduce it mod L and return the packed reduced result. See unpack_from_bytes() for more details on performed unpacking.